+++ src/pages/StartPage.vue
@@ -52,7 +52,7 @@ export default {
     border-radius: 0.5rem
     cursor: pointer
     transition: all 0.3s ease-in-out
-
     &:hover
       background-color: $primary-color-hover
+      
 </style>
